add methods up to the my_map
